requirements.txt
psycopg[binary]
pymupdf
openai
python-dotenv
pgvector

